# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: "Сервис помощи организации частных мероприятий"
  description: Тип - веб-приложение. Задачи - позволяет пользователям совместно планировать события, делить задачи, распределять расходы и отслеживать обязательства.
  version: 1.0.0
servers:
  - url: 'https'
security:
  - BearerAuth: []


# Добавленные теги
...
paths:
...
  /events/{event_id}/participants-list:
    get:
      tags:
        - Participants list
      operationId: getEventParticipantsList
      summary: Получить список участников мероприятия
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      description: Возвращает список всех участников мероприятия
      responses:
        '200':
          description: Participants list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '410':
          description: Не удалось просмотреть список участников
  /events/{event_id}/participants-list/add-participant:
    get:
      tags:
        - Participants list
      operationId: searchUsers
      summary: Найти 10 пользователей для добавления
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
        - in: query
          name: text
          required: true
          description: Текст из поисковой строки
          schema:
            type: string
        - in: query
          name: seq
          required: true
          description: Порядковый номер "пачки" участников
          schema:
            type: integer
      responses:
        '200':
          description: Пачка из 10 участников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDemo'
        '400':
          description: Неверные параметры запроса
        '404':
          description: Пользователь не найден
    post:
      tags:
        - Ignored
        - Participants list
      operationId: addParticipantsToParticipantsList
      summary: Добавить выбранных участников
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParticipantsRequest'
      responses:
        '200':
          description: Участники успешно добавлены
        '400':
          description: Неверный формат данных или некорректные логины
        '404':
          description: Событие не найдено
...
  /events/{event_id}/cost-list:
    get:
      tags:
        - Cost list
      operationId: getCostList
      summary: Получить список расходов
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cost list
          content:
            application/json:
              schema:
                type: object
                properties:
                  cost-allocation-list:
                    type: array
                    items:
                      $ref: '#/components/schemas/CostAllocationListItem'
                  expenses-existence: #смотрим по таблице покупок с таким event_id
                    type: boolean
        '404':
          description: Мероприятие с данным идентификатором не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventNotExistResponse'
  /events/{event_id}/cost-list/{purchase_id}/participants:
    get:
      tags:
        - Cost list
      operationId: getParticipantsForPurchaseFromCostList
      summary: Получить список участников для данной покупки
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
        - in: path
          name: purchase_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Participants list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDemo'
        '404':
          description: Мероприятие с данным идентификатором не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventNotExistResponse'
  /events/{event_id}/purchases-list/{purchase_id}/get-receipt:
    get:
      tags:
        - Ignored
        - Cost list
        - Cost allocation list
        - My purchases
      operationId: getReceiptsForPurchase
      summary: Получить чек покупки
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
        - in: path
          name: purchase_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Фото чека покупки
          content:
            multipart/form-data:
              schema:
                $ref: '#/components/schemas/ReceiptList'
              encoding:
                files:
                  contentType: image/jpeg
                  explode: true
        '404':
          description: Мероприятие с данным идентификатором не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventNotExistResponse'
  /events/{event_id}/cost-list-personal:
    get:
      tags:
        - Cost list
      operationId: getPersonalCostList
      summary: Получить список расходов по текущему пользователю
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cost list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostAllocationListItem'
        '404':
          description: Мероприятие с данным идентификатором не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventNotExistResponse'


  # my purchases list
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ReceiptList:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
          description: Массив фото чеков
    AddParticipantsRequest:
      type: array
      items:
        type: string
        minLength: 3
        maxLength: 20
        example: main_player
...
    EventNotExistResponse:
      type: object
      properties:
        error:
          type: string
          description: Подробное описание ошибки
...
    User:
      type: object
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 20
          example: main_player
        role_name:
          type: string
          example: создатель
        password:
          type: string
          minLength: 8
          maxLength: 20
          example: An1@!#fhv11
        email:
          type: string
          minLength: 1
          maxLength: 254
          example: poposh01@mail.ru
        name:
          type: string
          minLength: 1
          maxLength: 20
          example: Алина
        surname:
          type: string
          minLength: 1
          maxLength: 50
          example: Смирнов
        comment_money_transfer:
          type: string
          maxLength: 254
          example: "На сбер: 89996362576"
      required:
        - login
        - role_name
        - email
        - name
        - surname
...
    UserDemo:
      type: object
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 20
          example: main_player
        email:
          type: string
          minLength: 1
          maxLength: 254
          example: poposh01@mail.ru
        name:
          type: string
          minLength: 1
          maxLength: 20
          example: Алина
        surname:
          type: string
          minLength: 1
          maxLength: 50
          example: Смирнов
      required:
        - login
        - email
        - name
        - surname


    # Задача
...
    CostAllocationListItem:
      type: object
      properties:
        purchase_id:
          type: integer
          example: 1
        purchase_name:
          type: string
          minLength: 1
          maxLength: 50
          example: Огурцы
        responsible_name:
          type: string
        responsible_surname:
          type: string
        cost:
          type: number
          format: float
          example: 123.99
        hasReceipt:
          type: boolean
        countParticipants:
          type: integer
          example: 2
      required:
        - purchase_id
        - purchase_name
        - cost
        - hasReceipt

    # Чек
...